# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IUT Cafeteria API
  description: |
    API for the IUT Cafeteria ordering system.

    ## Authentication
    All endpoints (except /login, /register, /health) require JWT authentication.
    Pass the token via:
    - `Authorization: Bearer <token>` header
    - `token` httpOnly cookie

    ## Correlation IDs
    All requests are assigned an `x-correlation-id` header for distributed tracing.
  version: 1.0.0
  contact:
    name: IUT Cafeteria Team
    email: admin@iut-cafeteria.app
servers:
  - url: http://localhost:3000
    description: Order Gateway (Development)
  - url: http://localhost:3001
    description: Identity Provider (Development)
  - url: http://localhost:3002
    description: Stock Service (Development)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    Transaction:
      type: object
      properties:
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
          format: float
        balance_after:
          type: number
          format: float
        description:
          type: string
        created_at:
          type: string
          format: date-time
...
paths:
...
  /wallet/balance:
    get:
      tags: [Wallet]
      summary: Get wallet balance
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: Current balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: number
                    format: float
  /wallet/transactions:
    get:
      tags: [Wallet]
      summary: Get transaction history
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
  /wallet/topup:
    post:
      tags: [Wallet]
      summary: Top up wallet
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  minimum: 10
                method:
                  type: string
                  enum: [bkash, nagad, rocket, bank]
                  default: bkash
                reference:
                  type: string
      responses:
        '200':
          description: Top-up successful
  /wallet/emergency/status:
    get:
      tags: [Wallet]
      summary: Get emergency loan status
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: Emergency loan info
          content:
            application/json:
              schema:
                type: object
                properties:
                  outstanding:
                    type: number
                  available:
                    type: number
                  limit:
                    type: number
  /wallet/emergency/request:
    post:
      tags: [Wallet]
      summary: Request emergency balance
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  minimum: 10
                reason:
                  type: string
      responses:
        '200':
          description: Emergency balance added

  # ─── Cafeteria Tokens ─────────────────────────────────────────
...
  /admin/stock/{itemId}:
    put:
      tags: [Admin]
      summary: Update stock quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Stock updated
...
  /stock/{itemId}:
    get:
      tags: [Admin]
      summary: Get stock level
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Stock info
          content:
            application/json:
              schema:
                type: object
                properties:
                  itemId:
                    type: integer
                  quantity:
                    type: integer
                  source:
                    type: string
                    enum: [cache, database]
  /stock/{itemId}/decrement:
    post:
      tags: [Admin]
      summary: Decrement stock (internal)
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Stock decremented
        '409':
          description: Out of stock or conflict